backend/src/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.1.1
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.6
passlib==1.7.4
pydantic==2.5.2
pydantic-core==2.14.5
PyJWT==2.8.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
uvicorn==0.24.0.post1
